+++ s1tbx-op-coregistration/src/test/java/org/esa/nest/gpf/TestGCPSelectionOp.java
@@ -86,8 +86,8 @@ public class TestGCPSelectionOp {
         product.addBand(slvBand);
 
         // create lat/lon tie point grids
-        final float[] lat = new float[w * h];
-        final float[] lon = new float[w * h];
+        final double[] lat = new double[w * h];
+        final double[] lon = new double[w * h];
         for (int y = 0; y < h; y++) {
             for (int x = 0; x < w; x++) {
                 i = y * w + x;
