+++ s1tbx-op-sar-processing/src/main/java/org/esa/nest/gpf/geometric/SARSimulationOp.java
@@ -570,7 +570,7 @@ public final class SARSimulationOp extends Operator {
                         if (saveZeroHeightSimulation) {
                             alt = 1;
                         } else {
-                            geoPos.setLocation((float) lat, (float) lon);
+                            geoPos.setLocation(lat, lon);
                             alt = dem.getElevation(geoPos);
                             if (alt == demNoDataValue)
                                 continue;
@@ -693,7 +693,7 @@ public final class SARSimulationOp extends Operator {
                             double[] latlon = jOrbit.lp2ell(new Point(x + 0.5, y + 0.5), meta);
                             lat = latlon[0] * MathUtils.RTOD;
                             lon = latlon[1] * MathUtils.RTOD;
-                            alt = dem.getElevation(new GeoPos((float) lat, (float) lon));
+                            alt = dem.getElevation(new GeoPos(lat, lon));
                         }
 
                         if (!getPosition(lat, lon, alt, x0, y0, w, h, posData))
